// frontend/src/types/common/dialogs.types.ts
interface DialogExtraOptions {
  mode?: "click" | "selection";
  count?: number;
  maxWidth?: MuiDialogProps["maxWidth"];
  hasButtons?: boolean;
  isSingleton?: boolean;
}
// ...
export interface FormDialogProps
  extends FormButtonsProps,
    Omit<MuiDialogProps, "onSubmit" | "open">,
    DialogExtraOptions {
  open?: boolean;
  title?: string;
  children?: React.ReactNode;
}

// form
export interface FormButtonsProps {
  onSubmit?: (e: BaseSyntheticEvent) => void;
  onCancel?: () => void;
  cancelButtonProps?: ButtonProps;
  confirmButtonProps?: ButtonProps;
}

export type ComponentFormProps<T> = FormButtonsProps & {
  data: T | null;
  onError?: () => void;
  onSuccess?: () => void;
  Wrapper?: React.FC<FormDialogProps>;
  WrapperProps?: Partial<FormDialogProps> & Partial<FormButtonsProps>;
};